Experimental: router structure
