+++ config/locales/cy.yml
@@ -15,7 +15,6 @@ global:
     ukWide: DU gyfan
   nav:
     home: Hafan
-    toplevel: Hafan
     logIn: Mewngofnodi
     logOut: Allgofnodi
     myApplications: Fy ngheisiadau
