+++ config/locales/en.yml
@@ -15,7 +15,6 @@ global:
     ukWide: UK-wide
   nav:
     home: Home
-    toplevel: Home
     logIn: Log in
     logOut: Log out
     myApplications: My applications
