+++ controllers/about/our-people/views/profiles.njk
@@ -50,7 +50,7 @@
                     {# @TODO: Should we share this component? #}
                     <nav class="nested-menu">
                         <h2 class="nested-menu__section t4">
-                            <a class="nested-menu__section-link" href="./">
+                            <a class="nested-menu__section-link" href="../../../..">
                                 {{ sectionTitle }}
                             </a>
                         </h2>
