+++ controllers/archived.js
@@ -4,6 +4,38 @@ const router = express.Router();
 
 const { buildArchiveUrl, makeWelsh } = require('../common/urls');
 
+const aliases = require('./aliases');
+
+/**
+ * Handle Aliases
+ */
+aliases.forEach((redirect) => {
+    router.get(redirect.from, (req, res) => res.redirect(301, redirect.to));
+});
+
+/**
+ * Handle legacy programme pages as wildcards
+ * (eg. redirect them to /funding/programmes/<slug>)
+ */
+router.get('/:region?/global-content/programmes/:country/:slug', (req, res) => {
+    const locale = req.params.region === 'welsh' ? '/welsh' : '';
+    res.redirect(301, `${locale}/funding/programmes/${req.params.slug}`);
+});
+
+/**
+ * Handle migrated A Better Start child pages as wildcards
+ * (eg. redirect them to /funding/publications/a-better-start/<slug>)
+ */
+router.get(
+    '/funding/strategic-investments/a-better-start/:slug',
+    (req, res) => {
+        res.redirect(
+            301,
+            `/funding/publications/a-better-start/${req.params.slug}`
+        );
+    }
+);
+
 /**
  * Archived Routes
  * Paths in this array will be redirected to the National Archives.
