+++ controllers/funding/grants/index.js
@@ -14,9 +14,9 @@ const {
     injectCopy,
     injectHeroImage,
     setHeroLocals,
-} = require('../../common/inject-content');
-const { sMaxAge } = require('../../common/cached');
-const contentApi = require('../../common/content-api');
+} = require('../../../common/inject-content');
+const { sMaxAge } = require('../../../common/cached');
+const contentApi = require('../../../common/content-api');
 
 const grantsService = require('./grants-service');
