+++ controllers/funding/materials/views/materials.njk
@@ -2,7 +2,7 @@
 {% from "components/icons.njk" import iconClose %}
 {% from "./materials-fields.njk" import textInput, radios with context %}
 
-{% extends "controllers/common/views/cms-page.njk" %}
+{% extends "../../../common/views/cms-page.njk" %}
 
 {% macro textField(field) %}
     {{ textInput(
