+++ controllers/funding/strategic-investments/index.js
@@ -6,10 +6,10 @@ const compact = require('lodash/compact');
 const {
     injectListingContent,
     setCommonLocals,
-} = require('../../common/inject-content');
-const contentApi = require('../../common/content-api');
+} = require('../../../common/inject-content');
+const contentApi = require('../../../common/content-api');
 
-const { renderFlexibleContentChild } = require('../common');
+const { renderFlexibleContentChild } = require('../../common');
 
 const router = express.Router();
