+++ controllers/funding/strategic-investments/views/strategic-investments.njk
@@ -1,7 +1,7 @@
 {% from "components/hero.njk" import hero with context %}
 {% from "components/promo-card/macro.njk" import promoCard %}
 
-{% extends "controllers/common/views/information-page.njk" %}
+{% extends "../../../common/views/information-page.njk" %}
 
 {% block customContent %}
     {% if strategicProgrammes.length > 0 %}
