+++ controllers/funding/strategic-investments/views/strategic-programme.njk
