+++ server.js
@@ -1,7 +1,6 @@
 'use strict';
 const path = require('path');
 const express = require('express');
-const moment = require('moment');
 const i18n = require('i18n-2');
 const yaml = require('js-yaml');
 const nunjucks = require('nunjucks');
@@ -23,8 +22,7 @@ const { defaultMaxAge, noStore } = require('./common/cached');
 const cspDirectives = require('./common/csp-directives');
 const logger = require('./common/logger').child({ service: 'server' });
 
-const aliases = require('./controllers/aliases');
-const routes = require('./controllers/routes');
+const { basicContent } = require('./controllers/common');
 const { renderError, renderNotFound } = require('./controllers/errors');
 
 /**
@@ -61,42 +59,17 @@ i18n.expressBind(app, {
 });
 
 /**
- * Old domain redirect
- * Redirect requests from apply.tnlcommunityfund.org.uk
+ * Domain redirects
+ * Redirect requests from any non-canonical domains
  */
 app.use(require('./controllers/domain-redirect'));
 
 /**
- * Status endpoint
- * Mount first
+ * Metadata endpoints
+ * Mount first before any common middleware
  */
-const LAUNCH_DATE = moment();
-app.get('/status', (req, res) => {
-    res.setHeader('Access-Control-Allow-Origin', '*');
-    res.setHeader(
-        'Access-Control-Allow-Headers',
-        'Origin, X-Requested-With, Content-Type, Accept'
-    );
-    res.setHeader('Cache-Control', 'no-store,no-cache,max-age=0');
-
-    res.json({
-        APP_ENV: appData.environment,
-        DEPLOY_ID: appData.deployId,
-        COMMIT_ID: appData.commitId,
-        BUILD_NUMBER: appData.buildNumber,
-        START_DATE: LAUNCH_DATE.format('dddd, MMMM Do YYYY, h:mm:ss a'),
-        UPTIME: LAUNCH_DATE.toNow(true),
-    });
-});
-
-/**
- * Robots.txt
- */
+app.get('/status', require('./controllers/status'));
 app.get('/robots.txt', noStore, require('./controllers/robots'));
-
-/**
- * Site-map
- */
 app.use('/sitemap.xml', require('./controllers/sitemap'));
 
 /**
@@ -187,56 +160,68 @@ app.use([
 app.use('/api', require('./controllers/api'));
 app.use('/tools', require('./controllers/tools'));
 app.use('/patterns', require('./controllers/pattern-library'));
+app.use('/search', require('./controllers/search'));
 
 /**
- * Handle Aliases
- */
-aliases.forEach((redirect) => {
-    app.get(redirect.from, (req, res) => res.redirect(301, redirect.to));
-});
-
-/**
- * Handle legacy programme pages as wildcards
- * (eg. redirect them to /funding/programmes/<slug>)
- */
-app.get('/:region?/global-content/programmes/:country/:slug', (req, res) => {
-    const locale = req.params.region === 'welsh' ? '/welsh' : '';
-    res.redirect(301, `${locale}/funding/programmes/${req.params.slug}`);
-});
-
-/**
- * Handle migrated A Better Start child pages as wildcards
- * (eg. redirect them to /funding/publications/a-better-start/<slug>)
- */
-app.get('/funding/strategic-investments/a-better-start/:slug', (req, res) => {
-    res.redirect(
-        301,
-        `/funding/publications/a-better-start/${req.params.slug}`
-    );
-});
-
-/**
- * Archived paths
- * Handles archived pages (eg. redirect to National Archives)
- * and legacy files (eg. show a message about removed documents)
+ * Handle archived paths first:
+ * - Legacy redirects and aliases
+ * - National Archives content
+ * - Archived media
  */
 app.use('/', require('./controllers/archived'));
 
 /**
- * Initialise section routes
- * - Creates a new router for each section
- * - Apply shared middleware
- * - Apply section specific controller logic
- * - Add common routing (for static/fully-CMS powered pages)
+ * Initialise section routers
  */
-forEach(routes, function (section, sectionId) {
-    const router = express.Router();
+const sections = {
+    home: {
+        path: '/',
+        router: require('./controllers/home'),
+    },
+    funding: {
+        path: '/funding',
+        router: require('./controllers/funding'),
+    },
+    insights: {
+        path: '/insights',
+        router: require('./controllers/insights'),
+    },
+    contact: {
+        path: '/contact',
+        router: basicContent(),
+    },
+    about: {
+        path: '/about',
+        router: require('./controllers/about'),
+    },
+    updates: {
+        path: '/news',
+        router: require('./controllers/updates'),
+    },
+    jobs: {
+        path: '/jobs*',
+        router: basicContent(),
+    },
+    data: {
+        path: '/data',
+        router: require('./controllers/data'),
+    },
+    user: {
+        path: '/user',
+        router: require('./controllers/user'),
+    },
+    apply: {
+        path: '/apply',
+        router: require('./controllers/apply'),
+    },
+};
 
+forEach(sections, function (section, sectionId) {
     /**
      * Add section locals
      * Used for determining top-level section for navigation and breadcrumbs
      */
-    router.use(function (req, res, next) {
+    function sectionLocals(req, res, next) {
         const sectionTitle = req.i18n.__(`global.nav.${sectionId}`);
         const sectionUrl = localify(req.i18n.getLocale())(req.baseUrl || '/');
 
@@ -249,20 +234,13 @@ forEach(routes, function (section, sectionId) {
         res.locals.breadcrumbs = [{ label: sectionTitle, url: sectionUrl }];
 
         next();
-    });
+    }
 
-    /**
-     * Mount page router
-     */
-    section.pages.forEach(function (page) {
-        router.use(page.path, page.router);
-    });
-
-    /**
-     * Mount section router
-     */
-    app.use(section.path, router);
-    app.use(makeWelsh(section.path), router);
+    app.use(
+        [section.path, makeWelsh(section.path)],
+        sectionLocals,
+        section.router
+    );
 });
 
 /**
